Move _new_token function into token_flow
